fix: track generated json files
These files are necessary to build the parser from source with
`tree-sitter build` (which nvim-treesitter is transitioning to).

You could also drop the parser.c (which is huge) from the repo and
require generating before building (as a few grammars do), but your
grammar is very complex and takes too long to generate.